renamed react/jsx files from .js -> .jsx
